feat(reputation_base): create reusable base class for reputation extraction, use BaseReputationExtractor for consistency
